+++ packages/core/src/gh/_super/types.ts
@@ -85,6 +85,10 @@ export type GitHubOpts = {
 			 */
 			path    : string
 			/**
+			 * URL Content
+			 */
+			url     : string | undefined
+			/**
 			 * File Content
 			 */
 			content : Content