+++ packages/builder/CHANGELOG.md
@@ -1,3 +1,9 @@
 # @collectium/builder
 
+## 0.3.0
+
+### Minor Changes
+
+- Change content response type, add url to content response, fix some issues
+
 ## 0.2.3