+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @collectium/core
 
+## 0.3.0
+
+### Minor Changes
+
+- Change content response type, add url to content response, fix some issues
+
 ## 0.2.3
 
 ### Patch Changes