+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,18 @@
 # collectium
 
+## 0.3.0
+
+### Minor Changes
+
+- Change content response type, add url to content response, fix some issues
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @collectium/check@1.0.0
+  - @collectium/core@1.0.0
+  - @collectium/cli@1.0.0
+
 ## 0.2.3
 
 ### Patch Changes